separated backend config from health monitoring
