+++ src/SimpleL7Proxy/Backend/IBackendHostHealthCollection.cs
@@ -1,0 +1,14 @@
+﻿using Microsoft.Extensions.Logging;
+using Microsoft.Extensions.Options;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace SimpleL7Proxy.Backend;
+
+public interface IBackendHostHealthCollection
+{
+  List<BackendHostHealth> Hosts { get; }
+}
