--- src/SimpleL7Proxy/IBackendService.cs
@@ -1,15 +1,0 @@
-namespace SimpleL7Proxy;
-
-public interface IBackendService
-{
-    void Start();
-    public List<BackendHost> GetActiveHosts();
-
-    public int ActiveHostCount();
-
-    public Task WaitForStartup(int timeout);
-    public string HostStatus { get; }
-    public void TrackStatus(int code);
-    public bool CheckFailedStatus();
-    public string OAuth2Token();
-}
